addition of only internet for note details access

+++ app/src/main/java/com/example/finalproject/repositories/NotesRepository.kt
@@ -1,12 +1,19 @@
 package com.example.finalproject.repositories
 
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
+import android.os.Build
+import com.example.finalproject.SharedPreferencesUserDataStore
 import com.example.finalproject.api.ApiService
 import com.example.finalproject.models.Note
 import com.example.finalproject.room.NoteDao
 import kotlinx.coroutines.flow.flow
 
 class NotesRepository(private val noteDao: NoteDao,
-                      private val apiService: ApiService, private val user_id: String?
+                      private val apiService: ApiService,
+                      val userDataStore: SharedPreferencesUserDataStore,
+                        val context: Context
 ) {
 
     val notes = noteDao.getAllNotes()
@@ -47,17 +54,39 @@ class NotesRepository(private val noteDao: NoteDao,
     }
 
     fun getAll() = flow {
-        // if hay internet has esto
-        val result = apiService.getNotes(user_id)
-        if (result.isSuccessful && result.body() != null) {
-            noteDao.deleteAll()
-            noteDao.insertAll(result.body()!!)
-            emit(true)
+        if (isInternetAvailable(context)) {
+            val result = apiService.getNotes(userDataStore.getUserId())
+            if (result.isSuccessful && result.body() != null) {
+                noteDao.deleteAll()
+                noteDao.insertAll(result.body()!!)
+                emit(true)
+            } else {
+                emit(false)
+            }
         } else {
             emit(false)
         }
-        // else emit(false o errorCode de no internet)
     }
+
+    private fun isInternetAvailable(context: Context): Boolean {
+        val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            val network = connectivityManager.activeNetwork ?: return false
+            val activeNetwork = connectivityManager.getNetworkCapabilities(network) ?: return false
+            return when {
+                activeNetwork.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> true
+                activeNetwork.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> true
+                activeNetwork.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET) -> true
+                else -> false
+            }
+        } else {
+            @Suppress("DEPRECATION")
+            val networkInfo = connectivityManager.activeNetworkInfo ?: return false
+            @Suppress("DEPRECATION")
+            return networkInfo.isConnected
+        }
+    }
+
 
     suspend fun insertToApi(note: Note) {
         try {
